+++ src/app/services/user/auth.service.ts
@@ -1,20 +1,17 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { catchError, firstValueFrom } from 'rxjs';
+import { catchError, firstValueFrom, Observable } from 'rxjs';
 import { environment } from '../../../environments/environment';
 import { BaseService } from '../base/base.service';
 import { NotificationsService } from 'angular2-notifications';
 import { RegisterRequest } from '../../modules/register/components/register/register.model';
-import { errors as userErrors } from './user.errors';
+import { errors as userErrors } from './auth.errors';
 
 @Injectable({
   providedIn: 'root'
 })
-export class UserService extends BaseService {
+export class AuthService extends BaseService {
 
-  /**
-   * Waiting for homie when he adds that prefix 🤘
-   */
   readonly baseResource: string = 'auth';
 
   constructor(
