+++ src/app/shared/models/auth-scheme.model.ts
@@ -1,0 +1,1 @@
+export type AuthScheme = 'Bearer';
